filename fixed, configured routes properly
