--- postconfig.js
@@ -1,7 +1,0 @@
-export default {
-    plugins: {
-      tailwindcss: {},
-      autoprefixer: {},
-    },
-  }
-
