+++ src/main.jsx
@@ -1,10 +1,23 @@
+import "./index.css"
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App.jsx'
-import './index.css'
+import Home from "./routes/home.jsx"
+import Login from "./routes/login.jsx"
+import {createBrowserRouter, RouterProvider} from "react-router-dom"
+
+const router = createBrowserRouter([
+  {
+    path: "/", 
+    element: <Home />
+  },
+  {
+    path: "/login",
+    element: <Login />
+  }
+])
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <App />
+    <RouterProvider  router={router} />
   </React.StrictMode>,
 )
