+++ tailwind.config.js
@@ -1,22 +1,9 @@
-// /** @type {import('tailwindcss').Config} */
-// export default {
-//   content: ["./src/**/*.jsx"] 
-//   @tailwind base;
-//   @tailwind components;
-//   @tailwind utilities; 
-//   theme: {
-//     extend: {},
-//   },
-//   plugins: [],
-// }
-
-module.exports = {
-  content: [
-    "./src/**/*.{js,jsx}",
-  ],
-
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: ["./src/**/*.jsx"],
   theme: {
     extend: {},
   },
   plugins: [],
 }
+
